PYweb/requirements.txt
numpy
opencv-python
Pillow
scikit-learn
Flask
Flask-Cors
pywebview
